backend/requirements.txt
asgiref==3.11.0
beautifulsoup4==4.14.2
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
idna==3.11
pillow==12.0.0
openpyxl==3.1.5
PyJWT==2.10.1
requests==2.32.5
soupsieve==2.8
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.11.0
